# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API GoodMealAPP",
        "version": "1.0"
    },
    "paths": {
...
        "/api/shops": {
            "get": {
                "tags": [
                    "Shops"
                ],
                "summary": "Get list of shops",
                "description": "Returns list of shops",
                "operationId": "getShopsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Shops"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Shops"
                ],
                "summary": "Store new shop",
                "description": "Returns shop data",
                "operationId": "storeShop",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "image",
                                    "options",
                                    "address",
                                    "lon",
                                    "lat"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nuevo shop"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripcion de nuevo shop"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "https://via.placeholder.com/640x480.png/005522?text=alias"
                                    },
                                    "options": {
                                        "type": "string",
                                        "example": "pick-up | delivery | both"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Lorem ipsum dolor sit, amet consectetur adipisicing elit. Deserunt provident recusandae laboriosam rem, ratione id inventore tempora"
                                    },
                                    "lon": {
                                        "type": "string",
                                        "example": "-66.83921109896633"
                                    },
                                    "lat": {
                                        "type": "string",
                                        "example": "10.49215828524248"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Shops"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/shops/{shop}": {
            "put": {
                "tags": [
                    "Shops"
                ],
                "summary": "Update existing shop",
                "description": "Returns updated shop data",
                "operationId": "updateShop",
                "parameters": [
                    {
                        "name": "shop",
                        "in": "path",
                        "description": "Shop id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "image",
                                    "options",
                                    "address",
                                    "lon",
                                    "lat"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nuevo shop"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripcion de nuevo shop"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "https://via.placeholder.com/640x480.png/005522?text=alias"
                                    },
                                    "options": {
                                        "type": "string",
                                        "example": "pick-up | delivery | both"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Lorem ipsum dolor sit, amet consectetur adipisicing elit. Deserunt provident recusandae laboriosam rem, ratione id inventore tempora"
                                    },
                                    "lon": {
                                        "type": "string",
                                        "example": "-66.83921109896633"
                                    },
                                    "lat": {
                                        "type": "string",
                                        "example": "10.49215828524248"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Shops"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Shops"
                ],
                "summary": "Delete a Shop",
                "description": "Delete a Shop",
                "operationId": "destroyShop",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Shop--id"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Shop not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Shop": {
                "title": "Shop",
                "description": "Shop model",
                "required": [],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id of Shop",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "name of Shop",
                        "type": "string",
                        "example": "Nuevo shop"
                    },
                    "description": {
                        "title": "description",
                        "description": "description of Shop",
                        "type": "string",
                        "example": "Descripcion de nuevo shop"
                    },
                    "image": {
                        "title": "image",
                        "description": "image of Shop",
                        "type": "string",
                        "example": "https://via.placeholder.com/640x480.png/005522?text=alias"
                    },
                    "options": {
                        "title": "options",
                        "description": "options of Shop",
                        "type": "string",
                        "example": "pick-up || delivery || both"
                    },
                    "address": {
                        "title": "address",
                        "description": "address of Shop",
                        "type": "string",
                        "example": "Lorem ipsum dolor sit, amet consectetur adipisicing elit. Deserunt provident recusandae laboriosam rem, ratione id inventore tempora"
                    },
                    "lon": {
                        "title": "lon",
                        "description": "lon of Shop",
                        "type": "string",
                        "example": "-66.83921109896633"
                    },
                    "lat": {
                        "title": "lat",
                        "description": "lat of Shop",
                        "type": "string",
                        "example": "10.49215828524248"
                    },
                    "created_at": {
                        "title": "created_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-12-19T02:41:42.336Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-12-19T02:41:42.336Z"
                    }
                },
                "type": "object"
            },
            "Shops": {
                "title": "Shops",
                "properties": {
                    "data": {
                        "title": "data",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Shop"
                        }
                    }
                },
                "type": "object"
            }
        },
        "parameters": {
...
            "Shop--id": {
                "name": "shop",
                "in": "path",
                "description": "Id of Shop",
                "required": true,
                "schema": {
                    "type": "integer",
                    "example": "1"
                }
            }
        }
    }
}